Short to float conversions

+++ TR5Main/Objects/TR2/Vehicles/boat_info.h
@@ -2,9 +2,9 @@
 
 struct SpeedBoatInfo
 {
-	int TurnRate;
-	short LeanAngle;
-	short ExtraRotation;
+	float TurnRate;
+	float LeanAngle;
+	float ExtraRotation;
 
 	int LeftVerticalVelocity;
 	int RightVerticalVelocity;
